finish lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,8 +4,8 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double)top / bottom; // add (double)
+        int result = (int)Math.round(quotient); // add (int)
         return result;
     }
